requirements.txt
django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9
djangorestframework==3.16.0
markdown==3.8
django-filter==24.1
django-multiselectfield==0.1.13 